strategies/test-driven-development/requirements.txt
python-chess>=1.0
pytest>=7.0
